Improve performance, refactor utility roles
I've greatly improved the optimization of the importer. The tradeoff for this is now the importer no-longer automatically extracts materials from models when imported, but this isn't necessarily a bad thing. I want to bring it back later, but for now, in the interest of keeping it practical in speed, I'm going to leave it be.

I've also laid the groundworks for the future fluid model creator. I've done a lot of reading on ADT fluids, and have a method more or less in mind.