Update M2Utility.cs
I had made these objects their own unlinked prefabs in order to completely separate concerns with the model imports, but in retrospect, this adds a lot more frustrating complexity when you extract materials from a model to modify them. As such, going back to the original approach.

The long term goal is to make this more configurable for the end user, so they can have it any way they want it.