Ensure parsing floats works for cultures with different decimal separator (#2)
I think this is the way to fix it at least. It now correctly parses these american files in my country at least where comma is normally expected as the decimal separator. I haven't tested on an american computer to verify I didn't just break it there instead :p And I'm a noob at these things. This is just based on Googling and trial and error.