Add explicit Newtonsoft.Json dependency
We depend on this package, but this dependency wasn't explicitely defined so if the package is installed in a project not already containing Newtonsoft.Json then it will throw errors.

Note: I don't know how to add minimum dependencies in the Unity package manager, or if that is even supported. Like there's no specific reason I required version 3.2.1 of this package except for the fact that it was the latest version at the time. If I required 3.0.0 or 1.0.0 or whatever, then Unity would not fetch the latest version automatically.